survey_form/requirements.txt
asgiref==3.5.2
black==22.6.0
click==8.1.3
Django==4.1
django-extensions==3.2.0
djangorestframework==3.13.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pydot==1.4.2
pygraphviz==1.10
pyparsing==3.0.9
pytz==2022.2.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
